// src/theme/generated/extensions.ts
export const EXTENSION_THEME_COLORS: Record<string, ThemeInfo> = {
  "01e61500-862c-41cd-b653-48d2ef7b1882": {
    colors: {
      'editor.background': '#0B173A',
      'editor.foreground': '#F8F8F2',
      'titleBar.activeBackground': '#0B1633',
      'titleBar.inactiveBackground': '#0B1633',
      'statusBar.background': '#3EB7E5',
      'activityBar.background': '#0D193B',
      'activityBar.foreground': '#3EB7E5',
    },
    type: 'dark',
  },
// ...
};